added sparse progress to sparse handling after commit
needed because it is possible that sparse download is needed
