+++ versioncontrol/actions/git_repository_helper.py
@@ -115,3 +115,25 @@ class MergeProgress(vc.Progress):
 
     def canceled(self):
         return self.ap_progress.canceled
+    
+class SparseProgress(vc.Progress):
+    def __init__(self, progress: ap.Progress) -> None:
+        super().__init__()
+        self.ap_progress = progress
+
+    def update(self, operation_code: str, current_count: int, max_count: int, info_text: Optional[str] = None):
+        if operation_code == "downloading":
+            if info_text:
+                self.ap_progress.set_text(f"Downloading Files: {info_text}")
+            else:
+                self.ap_progress.set_text("Downloading Files")
+            self.ap_progress.report_progress(current_count / max_count)
+        elif operation_code == "updating":
+            self.ap_progress.set_text("Updating Files")
+            self.ap_progress.report_progress(current_count / max_count)
+        else:
+            self.ap_progress.set_text("Talking to Server")
+            self.ap_progress.stop_progress()
+
+    def canceled(self):
+        return self.ap_progress.canceled
